requirements.txt
git+https://github.com/deepseek-ai/Janus.git
torch
numpy
Pillow
aiohttp
transformers